# docs/swagger.yaml
basePath: /
definitions:
  types.APIResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
  types.BatchGetRecentPostsRequest:
    properties:
      end_time:
        description: 结束时间戳（可选）
        example: 1704153599
        type: integer
      platforms:
        description: 平台列表，最多10个平台
        items:
          properties:
            limit:
              description: 获取数量限制，默认10，最大100
              example: 10
              maximum: 100
              minimum: 1
              type: integer
            provider:
              description: 平台名称
              enum:
                - youtube
                - x
                - facebook
                - tiktok
                - instagram
              example: x
              type: string
          required:
            - provider
          type: object
        maxItems: 10
        minItems: 1
        type: array
      server_name:
        description: 服务名称
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      start_time:
        description: 开始时间戳（可选）
        example: 1704067199
        type: integer
      user_id:
        description: 用户ID
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - platforms
      - server_name
      - user_id
    type: object
  types.BatchGetRecentPostsResponse:
    properties:
      error_count:
        description: 查询失败的平台数量
        example: 1
        type: integer
      platforms:
        description: 各平台的帖子列表
        items:
          $ref: "#/definitions/types.PlatformPosts"
        type: array
      server_name:
        example: myapp
        type: string
      success_count:
        description: 成功查询的平台数量
        example: 3
        type: integer
      total_posts:
        description: 所有平台的总帖子数
        example: 25
        type: integer
      user_id:
        example: user123
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      request_id:
        type: string
    type: object
  types.GetRecentPostsRequest:
    properties:
      end_time:
        description: 结束时间戳（可选）
        example: 1704153599
        type: integer
      limit:
        description: 获取数量限制，默认10，最大100
        example: 10
        maximum: 100
        minimum: 1
        type: integer
      provider:
        description: 平台名称
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        description: 服务名称
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      start_time:
        description: 开始时间戳（可选）
        example: 1704067199
        type: integer
      user_id:
        description: 用户ID
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.GetRecentPostsResponse:
    properties:
      posts:
        description: 最近发布的帖子列表
        items:
          $ref: "#/definitions/types.Post"
        type: array
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      total:
        description: 总数量
        example: 10
        type: integer
      user_id:
        example: user123
        type: string
    type: object
...
  types.PlatformPosts:
    properties:
      error:
        description: 如果该平台查询失败，记录错误信息
        example: authentication failed
        type: string
      posts:
        description: 该平台的帖子列表
        items:
          $ref: "#/definitions/types.Post"
        type: array
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      total:
        description: 该平台的总数量
        example: 10
        type: integer
      user_id:
        example: user123
        type: string
    type: object
  types.Post:
    properties:
      content:
        description: 帖子内容
        example: Hello World!
        type: string
      created_at:
        description: 创建时间戳
        example: 1704067199
        type: integer
      description:
        description: 描述
        example: Post description
        type: string
      id:
        description: 帖子ID
        example: "1234567890"
        type: string
      media_type:
        description: 媒体类型：image, video, audio
        example: image
        type: string
      media_url:
        description: 媒体URL
        example: https://example.com/image.jpg
        type: string
      stats:
        allOf:
          - $ref: "#/definitions/types.StatsData"
        description: 统计信息
      title:
        description: 标题（YouTube等平台）
        example: My Post
        type: string
      updated_at:
        description: 更新时间戳
        example: 1704067199
        type: integer
      url:
        description: 帖子链接
        example: https://x.com/user/status/1234567890
        type: string
    type: object
...
  types.ShareRequest:
    properties:
      content:
        description: text content
        example: Hello World!
        maxLength: 280
        type: string
      description:
        example: This is a description
        maxLength: 500
        type: string
      media_url:
        description: url to media (backend should download & upload)
        example: https://example.com/image.jpg
        type: string
      privacy:
        enum:
          - public
          - private
          - unlisted
          - friends
          - followers
        example: public
        type: string
      provider:
        description: 平台名称 可选值：youtube x facebook tiktok instagram
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        description: 服务名称 必填
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      tags:
        example:
          - hello
          - world
        items:
          type: string
        maxItems: 10
        type: array
      title:
        example: My Post
        maxLength: 100
        type: string
      user_id:
        description: 用户ID 必填 同一服务名称下user_id唯一
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.ShareResponse:
    properties:
      content:
        example: "Hello from Social Platform! \U0001F680"
        type: string
      media_id:
        description: Tweet ID or post ID for status query
        example: "1234567890"
        type: string
      media_url:
        example: https://example.com/image.jpg
        type: string
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      tags:
        example:
          - social
          - oauth
          - test
        items:
          type: string
        type: array
      user_id:
        example: user123
        type: string
    type: object
...
  types.StatsData:
    properties:
      likes:
        example: 100
        type: integer
      replies:
        example: 25
        type: integer
      retweets:
        example: 50
        type: integer
      shares:
        example: 10
        type: integer
      views:
        example: 1000
        type: integer
    type: object
  types.StatsRequest:
    properties:
      media_id:
        example: "1234567890"
        maxLength: 100
        type: string
      provider:
        description: 平台名称 可选值：youtube x facebook tiktok instagram
        enum:
          - youtube
          - x
          - facebook
          - tiktok
          - instagram
        example: x
        type: string
      server_name:
        example: myapp
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        description: 用户ID 必填 同一服务名称下user_id唯一
        example: user123
        maxLength: 100
        minLength: 1
        type: string
    required:
      - provider
      - server_name
      - user_id
    type: object
  types.StatsResponse:
    properties:
      media_id:
        example: "1234567890"
        type: string
      provider:
        example: x
        type: string
      server_name:
        example: myapp
        type: string
      stats:
        $ref: "#/definitions/types.StatsData"
      user_id:
        example: user123
        type: string
    type: object
...
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 多平台社交媒体授权分享API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Platform API
  version: "1.0"
paths:
  /api/batch-recent-posts:
    post:
      consumes:
        - application/json
      description: 批量获取多个平台最近发布的内容列表，支持定时后驱
      parameters:
        - description: 批量获取最近发布内容请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.BatchGetRecentPostsRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.BatchGetRecentPostsResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 批量获取最近发布的内容
      tags:
        - 内容
  /api/recent-posts:
    post:
      consumes:
        - application/json
      description: 获取指定平台最近发布的内容列表
      parameters:
        - description: 获取最近发布内容请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.GetRecentPostsRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.GetRecentPostsResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 获取最近发布的内容
      tags:
        - 内容
  /api/share:
    post:
      consumes:
        - application/json
      description: 将内容分享到指定的社交媒体平台
      parameters:
        - description: 分享请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.ShareRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 分享成功
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.ShareResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 分享内容到社交媒体平台
      tags:
        - 分享
  /api/stats:
    post:
      consumes:
        - application/json
      description: 获取指定媒体内容在社交媒体平台上的统计信息
      parameters:
        - description: 统计请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/types.StatsRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 统计信息
          schema:
            allOf:
              - $ref: "#/definitions/types.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/types.StatsResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/types.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/types.ErrorResponse"
      summary: 获取社交媒体内容统计信息
      tags:
        - 统计
...
schemes:
  - http
  - https
swagger: "2.0"